bind the types before binding the constraints
