+++ test/sql/collate/test_unsupported_collations.test
@@ -6,11 +6,13 @@
 statement error
 CREATE TABLE collate_test(s VARCHAR COLLATE blabla)
 ----
+Catalog Error: Collation with name blabla does not exist!
 
 # non-varchar columns cannot have collations
 statement error
 CREATE TABLE collate_test(s INTEGER COLLATE blabla)
 ----
+Parser Error: Only VARCHAR columns can have collations!
 
 # we can combine multiple of the same collation
 statement ok
@@ -19,11 +21,14 @@ CREATE TABLE collate_test(s VARCHAR COLLATE NOACCENT.NOACCENT)
 statement error
 CREATE TABLE collate_test(s VARCHAR COLLATE 1)
 ----
+Parser Error: syntax error at or near "1"
 
 statement error
 CREATE TABLE collate_test(s VARCHAR COLLATE 'hello')
 ----
+Parser Error: syntax error at or near "'hello'"
 
 statement error
 PRAGMA default_collation='blabla'
 ----
+Catalog Error: Collation with name blabla does not exist!
